Fixed support of Reson 7k data (format 88) to work with MARUM AUV data. Includes changes to mb7kpreprocess.
Fixed 64 bit issues with Simrad, Reson 7k, and L3 XSE data
(formats 58, 59, 88, 94).

Fixed 64 bit issues with mbnavedit, mbclean, and mbneptune2esf.

Added buttons to MBedit to immediately go to the start or end of the data.

Fixed sidescan beamwidths used by the Edgetech Jstar format i/o
modules (formats 132 & 133).

Fixed memory management errors in MBgrdviz and mbfilter.

+++ trunk/src/mbview/mbgrdviz_callbacks.c
@@ -4263,7 +4263,7 @@ void do_mbgrdviz_open_region( Widget w, XtPointer client_data, XtPointer call_da
 	int	i, j, k, ksource;
 	
 	/* existing mbview instance */
-	int	instance_source;
+	size_t	instance_source;
 	struct mbview_struct *data_source;
 	char	button_name_source[MB_PATH_MAXLINE];
 	
@@ -4307,7 +4307,7 @@ void do_mbgrdviz_open_region( Widget w, XtPointer client_data, XtPointer call_da
 	float	*mbv_secondary_data;
    
     	/* get source mbview instance */
-	instance_source = (int) client_data;
+	instance_source = (size_t) client_data;
 /*fprintf(stderr,"Called do_mbgrdviz_open_region instance:%d\n", instance_source);*/
     
 	/* print input debug statements */
